% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_genes.R
\name{update_genes}
\alias{update_genes}
\title{Update Gene Symbols and Normalize Seurat Object}
\usage{
update_genes(
  Seurat_obj,
  assay = "RNA",
  return_seurat = TRUE,
  min.cells = 3,
  min.features = 200
)
}
\arguments{
\item{Seurat_obj}{A Seurat object. The RNA or any specified assay counts will be used.}

\item{assay}{The name of the assay from which to extract the counts matrix. Default is "RNA".}

\item{return_seurat}{Logical. If TRUE, a new Seurat object is returned with the updated gene names
and percentage features for mitochondrial and ribosomal genes. If FALSE, the function returns the
updated counts matrix. Default is TRUE.}

\item{min.cells}{Numeric. The minimum number of cells in which a gene must be detected to be
retained in the Seurat object. Default is 3.}

\item{min.features}{Numeric. The minimum number of features (genes) that must be detected in a
cell for that cell to be retained in the Seurat object. Default is 200.}
}
\value{
If \code{return_seurat} is TRUE, the function returns a new Seurat object
with updated gene names and two additional metadata columns: \code{percent.mt}, which indicates
the percentage of mitochondrial genes, and \code{percent.rb}, which shows the percentage
of ribosomal genes. If \code{return_seurat} is FALSE, the function returns the updated
counts matrix with corrected gene names.
}
\description{
This function updates gene names in a Seurat object using the latest gene symbol mapping
table from the HGNChelper package. It filters out low-expressed genes, updates gene names
to their approved symbols, and handles duplicated genes. Optionally, it can return
a new Seurat object with updated metadata, including mitochondrial and ribosomal gene
percentages.
}
\details{
This function leverages the HGNChelper package to ensure that all gene names
in a Seurat object match the currently approved HGNC symbols. It filters out genes
detected in fewer than \code{min.cells} cells and performs several operations to
handle duplicated gene symbols. For downstream analysis, it adds the proportion of
mitochondrial genes (if applicable) and ribosomal genes, accounting for both human
and mouse species.
}
\examples{
\dontrun{
# Example with a Seurat object
updated_seurat <- update_genes(Seurat_obj, assay = "RNA")
updated_matrix <- update_genes(Seurat_obj, assay = "RNA", return_seurat = FALSE)
}

}
\seealso{
\code{\link[Seurat]{CreateSeuratObject}}, \code{\link[Seurat]{PercentageFeatureSet}}, \code{\link[HGNChelper]{hgnc.table}}
}
